Refactoring: Moved data annotations to model package. Updated README with new features.
