+++ src/test/java/io/github/sskorol/model/Animal.java
@@ -1,6 +1,5 @@
 package io.github.sskorol.model;
 
-import io.github.sskorol.csv.Source;
 import lombok.Data;
 
 @Data
