+++ src/test/java/io/github/sskorol/model/CrimeRecord.java
@@ -1,7 +1,5 @@
 package io.github.sskorol.model;
 
-import io.github.sskorol.csv.Source;
-import io.github.sskorol.csv.FieldName;
 import lombok.Data;
 
 @Data
